Refactor to keep track of top-level source directories
