+++ src/main/kotlin/com/daveme/chocolateCakePHP/files.kt
@@ -36,51 +36,3 @@ fun findRelativeFile(dir: PsiDirectory?, childPath: String): VirtualFile? {
     val pathPartsArray = pathPartsList.dropLastWhile { it.isEmpty() }.toTypedArray()
     return VfsUtil.findRelativeFile(dir.virtualFile, *pathPartsArray)
 }
-
-fun templatePathToVirtualFile(
-    settings: Settings,
-    appOrPluginDir: PsiDirectory?,
-    controllerName: String,
-    controllerAction: String
-): VirtualFile? {
-    var relativeFile: VirtualFile? = null
-    if (settings.cake3Enabled) {
-        val underscored = controllerAction.camelCaseToUnderscore()
-        val cakeThreeTemplatePath = CakeThree.templatePath(settings, controllerName, underscored)
-        relativeFile = findRelativeFile(appOrPluginDir, cakeThreeTemplatePath)
-        if (relativeFile == null) {
-            val cakeFourTemplatePath = CakeFour.templatePath(settings, controllerName, underscored)
-            relativeFile = findRelativeFile(appOrPluginDir, cakeFourTemplatePath)
-        }
-    }
-    if (relativeFile == null) {
-        if (settings.cake2Enabled) {
-            val cakeTwoTemplatePath = CakeTwo.templatePath(settings, controllerName, controllerAction)
-            relativeFile = findRelativeFile(appOrPluginDir, cakeTwoTemplatePath)
-        }
-    }
-    return relativeFile
-}
-
-fun elementPathToVirtualFile(
-    settings: Settings,
-    pluginOrAppDir: PsiDirectory?,
-    elementPath: String
-): VirtualFile? {
-    var relativeFile: VirtualFile? = null
-    if (settings.cake3Enabled) {
-        val cakeThreeElementFilename = CakeThree.elementPath(settings, elementPath)
-        relativeFile = findRelativeFile(pluginOrAppDir, cakeThreeElementFilename)
-        if (relativeFile == null) {
-            val cakeFourElementFilename = CakeFour.elementPath(settings, elementPath)
-            relativeFile = findRelativeFile(pluginOrAppDir, cakeFourElementFilename)
-        }
-    }
-    if (relativeFile == null) {
-        if (settings.cake2Enabled) {
-            val cakeTwoElementFilename = CakeTwo.elementPath(settings, elementPath)
-            relativeFile = findRelativeFile(pluginOrAppDir, cakeTwoElementFilename)
-        }
-    }
-    return relativeFile
-}
