+++ src/main/kotlin/com/daveme/chocolateCakePHP/view/ElementGotoDeclarationHandler.kt
@@ -1,6 +1,8 @@
 package com.daveme.chocolateCakePHP.view
 
 import com.daveme.chocolateCakePHP.*
+import com.daveme.chocolateCakePHP.cake.elementPathToVirtualFile
+import com.daveme.chocolateCakePHP.cake.topSourceDirectoryFromFile
 import com.intellij.codeInsight.navigation.actions.GotoDeclarationHandler
 import com.intellij.openapi.actionSystem.DataContext
 import com.intellij.openapi.editor.Editor
@@ -32,8 +34,9 @@ class ElementGotoDeclarationHandler : GotoDeclarationHandler {
         }
         val containingFile = psiElement.containingFile
         val pluginOrAppDir = topSourceDirectoryFromFile(settings, containingFile)
+            ?: return PsiElement.EMPTY_ARRAY
         val relativeFile = elementPathToVirtualFile(settings, pluginOrAppDir, psiElement.text)
-                ?: return PsiElement.EMPTY_ARRAY
+            ?: return PsiElement.EMPTY_ARRAY
 
         val files = HashSet<VirtualFile>()
         files.add(relativeFile)
