+++ src/main/kotlin/com/daveme/chocolateCakePHP/view/TemplateGotoDeclarationHandler.kt
@@ -1,6 +1,8 @@
 package com.daveme.chocolateCakePHP.view
 
 import com.daveme.chocolateCakePHP.*
+import com.daveme.chocolateCakePHP.cake.templatePathToVirtualFile
+import com.daveme.chocolateCakePHP.cake.topSourceDirectoryFromFile
 import com.intellij.codeInsight.navigation.actions.GotoDeclarationHandler
 import com.intellij.openapi.actionSystem.DataContext
 import com.intellij.openapi.editor.Editor
@@ -36,8 +38,9 @@ class TemplateGotoDeclarationHandler : GotoDeclarationHandler {
         val controllerName = filename.controllerBaseName() ?: return PsiElement.EMPTY_ARRAY
 
         val pluginOrAppDir = topSourceDirectoryFromFile(settings, containingFile)
+            ?: return PsiElement.EMPTY_ARRAY
         val relativeFile = templatePathToVirtualFile(settings, pluginOrAppDir, controllerName, psiElement.text)
-                ?: return PsiElement.EMPTY_ARRAY
+            ?: return PsiElement.EMPTY_ARRAY
 
         val files = HashSet<VirtualFile>()
         files.add(relativeFile)
