+++ src/main/kotlin/com/daveme/chocolateCakePHP/view/ThisVariableInViewTypeProvider.kt
@@ -1,7 +1,7 @@
 package com.daveme.chocolateCakePHP.view
 
 import com.daveme.chocolateCakePHP.Settings
-import com.daveme.chocolateCakePHP.isCakeViewFile
+import com.daveme.chocolateCakePHP.cake.isCakeViewFile
 import com.daveme.chocolateCakePHP.viewType
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiElement
